# Speed up Utils.__listify and Utils.__pathify