--- a/elasticsearch-api/lib/elasticsearch/api/utils.rb
+++ b/elasticsearch-api/lib/elasticsearch/api/utils.rb
@@ -59,7 +59,7 @@ module Elasticsearch
       def __pathify(*segments)
         Array(segments).flatten.
           compact.
-          reject { |s| s.to_s =~ /^\s*$/ }.
+          reject { |s| s.squeeze(' ').size == 0 }.
           join('/').
           squeeze('/')
       end