##### Speed up `Elasticsearch::API::Utils.__listify`

Change `.map{ ... }.flatten` to `.flat_map` and
fetch escape option only once

```
    Calculating -------------------------------------
               __listify     67.427k (± 7.1%) i/s -      2.014M in  30.028225s
           __listify_new     78.107k (± 3.2%) i/s -      2.345M in  30.057810s

    Comparison:
           __listify_new:    78106.7 i/s
               __listify:    67426.5 i/s - 1.16x  slower

```


##### Speed up `Elasticsearch::API::Utils.__pathify`

Use String#squeeze instead of a Regexp to check for blank strings

```
    Calculating -------------------------------------
               __pathify    241.514k (± 4.4%) i/s -      7.246M in  30.067229s
           __pathify_new    277.492k (± 3.9%) i/s -      8.323M in  30.043878s

    Comparison:
           __pathify_new:   277492.5 i/s
               __pathify:   241514.4 i/s - 1.15x  slower
```

[Benchmark script](https://gist.github.com/jamescook/14d8f9c19fa731a73bd26af5e61f46f5)