@karmi 2017-10-05
Hi @jamescook, this looks like a great optimization! Apologies for not responding sooner.

I'm a bit on the fence regarding the `flat_map`, because it's only available since Ruby 1.9 (as far as I know), and the "core" gem is still compatible with Ruby 1.8. While there has been some discussion about whether it makes sense today, I'm somewhat reluctant to throw it away.

Based on the little reading I did, passing the `level` argument to (`Array#flatten`)[https://ruby-doc.org/core-1.8.7/Array.html#method-i-flatten] should achieve the same. Would you be so kind and run the benchmark script with it as well?

Regarding the check for empty string, the Regex is definitely an overkill there — would you be able to run the benchmark script with (`String#blank?`)[https://ruby-doc.org/core-1.8.7/String.html#method-i-empty-3F] if the performance benefit is observable? (I would bet it is.)