@jamescook 2017-10-08
@karmi `String#empty?` is indeed faster, a 1.72x speedup vs 1.15x. However, it doesn't preserve the original behavior regarding a string of multiple blanks, e.g. "   ".empty? returns false. If it's unnecessary to retain that behavior then String#empty? is definitely the better choice.

As for `flatten(1)` in place of `flat_map`, [it's not the same in terms of performance](https://github.com/JuanitoFatas/fast-ruby#enumerablemaparrayflatten-vs-enumerableflat_map-code). My local tests showed flatten(1) performed the same as flatten and slower than flat_map. From what I've read about flat_map, the performance boost is related to removing the need for an intermediate array.  The only workaround I can think of is a version guard like this:

```ruby
if RUBY_VERSION < '1.9'
 def __listify
   # ...
 end
else
  def __listify  # using flat_map
  # ...
  end
end
```
It would be nice if `bundle` ran successfully on 1.8 so I can run the tests, but it fails to install one of the dev dependencies (faraday).