@karmi 2018-01-14
Apologies for the delay, @jamescook! I have forgot about the patch altogether for a while, sorry.

Most importantly, I think time has come to drop the official support for 1.8 — you're correct that just installing gems (Faraday, Curb, ...) is difficult, we don't run the CI against this version anyway, it has been EOL-ed more than three years ago, ...

I've summarized the reasoning in 3ac633214173178e27971be73366a6ab27895fe8, moreover, due to the open nature of Ruby, it's not prohibitive for users on the old version to monkeypatch the library method or implement the new `Array#flat_map` method.

I've changed the check for empty string to `reject { |s| s.to_s.strip.empty? }` in e8e924d0a7441d38ea02e01fbb55ef2e47d70972, I think it's cleaner.

Once again, _many_ thanks for the patch, and for your patience. You should have pinged me sooner! :)