@karmi 2018-01-14
I've merged the commits into the `6.x` branch as well, so the improvements are picked up for the current version.

Travis CI passes both for [`master`](https://travis-ci.org/elastic/elasticsearch-ruby/builds/328775422) and [`6.x`](https://travis-ci.org/elastic/elasticsearch-ruby/builds/328778599).